+++ client/src/test/scala/stasis/test/specs/unit/client/ops/backup/stages/internal/PartitionedByteStringSourceSpec.scala
@@ -61,7 +61,7 @@ class PartitionedByteStringSourceSpec extends AsyncUnitSpec {
     val extended = new PartitionedByteStringSource(original)
 
     extended
-      .partition(withMaximumPartSize = maxSize)
+      .partition(withMaximumPartSize = maxSize.toLong)
       .mergeSubstreams
       .runWith(Sink.seq)
       .map { result =>
