+++ core/src/test/scala/stasis/test/specs/unit/core/persistence/staging/StagingStoreSpec.scala
@@ -269,7 +269,7 @@ class StagingStoreSpec extends AsyncUnitSpec with Eventually with BeforeAndAfter
 
   private val testManifest = Manifest(
     crate = Crate.generateId(),
-    size = testContent.size,
+    size = testContent.size.toLong,
     copies = 4,
     source = Node.generateId(),
     origin = Node.generateId()
